+++ sdk/agrifood/arm-agrifood/CHANGELOG.md
@@ -1,15 +1,63 @@
 # Release History
+    
+## 1.0.0-beta.5 (2022-12-12)
+    
+**Features**
 
-## 1.0.0-beta.4 (Unreleased)
+  - Added operation group Solutions
+  - Added operation group SolutionsDiscoverability
+  - Added operation Extensions.createOrUpdate
+  - Added Interface ApiProperties
+  - Added Interface ExtensionInstallationRequest
+  - Added Interface ExtensionsCreateOrUpdateOptionalParams
+  - Added Interface FarmBeatsSolution
+  - Added Interface FarmBeatsSolutionListResponse
+  - Added Interface FarmBeatsSolutionProperties
+  - Added Interface Insight
+  - Added Interface InsightAttachment
+  - Added Interface MarketplaceOfferDetails
+  - Added Interface Measure
+  - Added Interface ResourceParameter
+  - Added Interface Solution
+  - Added Interface SolutionEvaluatedOutput
+  - Added Interface SolutionInstallationRequest
+  - Added Interface SolutionListResponse
+  - Added Interface SolutionProperties
+  - Added Interface SolutionsCreateOrUpdateOptionalParams
+  - Added Interface SolutionsDeleteOptionalParams
+  - Added Interface SolutionsDiscoverabilityGetOptionalParams
+  - Added Interface SolutionsDiscoverabilityListNextOptionalParams
+  - Added Interface SolutionsDiscoverabilityListOptionalParams
+  - Added Interface SolutionsGetOptionalParams
+  - Added Interface SolutionsListNextOptionalParams
+  - Added Interface SolutionsListOptionalParams
+  - Added Type Alias ExtensionsCreateOrUpdateResponse
+  - Added Type Alias SolutionsCreateOrUpdateResponse
+  - Added Type Alias SolutionsDiscoverabilityGetResponse
+  - Added Type Alias SolutionsDiscoverabilityListNextResponse
+  - Added Type Alias SolutionsDiscoverabilityListResponse
+  - Added Type Alias SolutionsGetResponse
+  - Added Type Alias SolutionsListNextResponse
+  - Added Type Alias SolutionsListResponse
+  - Interface Extension has a new optional parameter additionalApiProperties
+  - Interface PrivateEndpointConnection has a new optional parameter groupIds
+  - Class AgriFoodMgmtClient has a new parameter solutionId
+  - Class AgriFoodMgmtClient has a new parameter solutions
+  - Class AgriFoodMgmtClient has a new parameter solutionsDiscoverability
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
+  - Removed operation Extensions.create
+  - Removed operation Extensions.update
+  - Class AgriFoodMgmtClient has a new signature
+    
+    
+## 1.0.0-beta.4 (2022-11-14)
+    
+**Features**
 
-### Bugs Fixed
-
-### Other Changes
-
+  - Bugs Fixed
+    
 ## 1.0.0-beta.3 (2022-10-08)
 
 **Bugs Fixed**